pyproject.toml
[tool.poetry]
name = "analisenoticias"
version = "0.1.0"
description = ""
authors = ["Gustavo Miranda <gustavocataldomiranda@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
isort = "^5.13.2"
blue = "^0.9.1"
pydocstyle = "^6.3.0"
pre-commit = "^3.8.0"
newsapi-python = "^0.2.7"
python-dotenv = "^1.0.1"
ipykernel = "^6.29.5"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
